worker/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
dapr==1.14.0
dapr-ext-fastapi==1.14.0
groq==0.12.0
sqlmodel==0.0.22
psycopg2-binary==2.9.10
python-dotenv==1.0.1
